requirements.txt
python-dotenv~=1.2.1

openai~=2.14.0

pillow~=12.1.0
pandas~=2.3.3
plotly~=6.5.0
pydantic~=2.12.5

fastapi~=0.128.0

langdetect~=1.0.9

# Repository
requests~=2.32.5
aiohttp~=3.13.3
openfoodfacts~=3.3.0

# Database
pymongo~=4.15.5

# Webapp framework
streamlit~=1.52.2
altair~=6.0.0

# Domain
google-genai~=1.60.0
python-dateutil~=2.9.0
tzlocal~=4.3.1